+++ README.md
@@ -38,7 +38,8 @@
   - 중복되는 역 이름, 노선 이름 검증 함수 통합 v
 - 리팩토링
   - 자바스크립트 컨벤션에 맞추어 validation depth줄이기 및 리팩토링 v
-  - css 검토
+  - 폴더 재구성 v
+  - html tag 검토 (인풋 밸류, 글자 크기)
 
 ## 🤦🏻‍♀️ 고민한 지점
