+++ src/index.js
@@ -1,7 +1,3 @@
 import main from "./main.js";
 
-export default function init() {
-  main();
-}
-
-init();
+main();
