+++ src/manager/handlers/lineHandlers.js
@@ -1,4 +1,4 @@
-import render from "../../components/render.js";
+import render from "../render.js";
 import app from "../../components/app.js";
 import { validateInput, validateStartAndEndStations } from "../validation/validation.js";
