+++ src/manager/render.js
